checkpoint: move bond polarizability to new crate
