+++ crates.d/rsp2-bond-polarizability.Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+!!CRATE-NAME-LINE
+version = "0.1.0"
+authors = ["Michael Lamparski <diagonaldevice@gmail.com>"]
+!!LICENSE-PERMISSIVE
+!!DEFAULT-EDITION
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+!!rsp2-array-types
+!!rsp2-structure
+!!rsp2-util-macros
+
+!!enum-map
+
+[features]
+!!NIGHTLY-FEATURE-LINE
