+++ src/math/bond-polarizability/Cargo.toml
@@ -1,0 +1,25 @@
+# DO NOT EDIT THIS FILE BY HAND!!
+# It has been autogenerated from crates.d/rsp2-bond-polarizability.Cargo.toml
+#
+# To update it, run './crates gen' in the workspace root
+
+[package]
+name = "rsp2-bond-polarizability"
+version = "0.1.0"
+authors = ["Michael Lamparski <diagonaldevice@gmail.com>"]
+license = "MIT OR Apache 2.0"
+edition = "2018"
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+rsp2-array-types = { path = "../../util/array-types" }
+rsp2-structure = { path = "../../structure" }
+rsp2-util-macros = { path = "../../util/macros" }
+
+enum-map = "0.4.1"
+
+[features]
+nightly = ["beta", "rsp2-array-types/nightly", "rsp2-structure/nightly", "rsp2-util-macros/nightly"]
+beta = ["rsp2-array-types/beta", "rsp2-structure/beta", "rsp2-util-macros/beta"]
